Changes as suggested by timgraham
* release text changes
* add trailing commas on list items
* split added tests into new test_uuid file
* use setUpTestData for tests with common data
