+++ docs/spelling_wordlist
@@ -578,7 +578,6 @@ postgis
 postgres
 Postgres
 postgresql
-PostgreSQL
 pq
 pre
 precisions
